feat: relocate-functionallong-replay-thin-event-helpers-001 - Keep long-lane replay thin-event behavior intact while relocating long-only helpers
